requirements.txt
Flask==2.3.3
qrcode==7.4.2
Pillow==10.0.0
pyperclip==1.8.2
